LayoutTests/platform/mac-snowleopard/accessibility/aria-list-and-listitem-expected.txt



This tests that the ARIA roles of list and listitem map correctly to Mac accessibility roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.childAtIndex(0).role is 'AXRole: AXList'
PASS obj.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(0).childAtIndex(1).role is 'AXRole: AXGroup'
PASS successfullyParsed is true

TEST COMPLETE
